+++ src/main.py
@@ -77,7 +77,7 @@ class Application():
 
         create_node_from_url_usecase = CreateNodeFromUrlUseCase(create_fragment_usecase, ressource_service)
 
-        node_orchestrator = NodeOrchestrator(node_service, fragment_service, spore_service, priority_service, ressource_service, create_node_from_url_usecase, node_view_builder)
+        node_orchestrator = NodeOrchestrator(node_service, fragment_service, spore_service, priority_service, ressource_service, create_node_from_url_usecase, node_view_builder, node_format_service)
         review_orchestrator = ReviewOrchestrator(user_service, node_service, review_service, node_view_builder)
 
         self.cleanup_service = CleanupService(node_service, collection_service, user_service)